Move common components to separate folder
