+++ src/app/components/common/progressBar.jsx
